Silly silly Hald thinks he can move the login function, no no
