+++ app/_layout.tsx
@@ -4,24 +4,19 @@ import {
   ThemeProvider,
 } from "@react-navigation/native";
 import { useFonts } from "expo-font";
-import { Stack, useRouter } from "expo-router";
+import { Stack } from "expo-router";
 import * as SplashScreen from "expo-splash-screen";
 import { StatusBar } from "expo-status-bar";
 import { useEffect } from "react";
 import "react-native-reanimated";
 
 import { useColorScheme } from "@/hooks/useColorScheme";
-import { UserProvider, useUser } from "@/lib/states/userState";
-import { deleteToken, getToken } from "@/lib/server/config";
-import { verify } from "@/lib/server/login";
-import { User } from "@/lib/API/schemas/User";
-import { get as getUser } from "@/lib/server/user";
+import { UserProvider } from "@/lib/states/userState";
 
 // Prevent the splash screen from auto-hiding before asset loading is complete.
 SplashScreen.preventAutoHideAsync();
 
 export default function RootLayout() {
-  const router = useRouter();
   const colorScheme = useColorScheme();
   const [loaded] = useFonts({
     SpaceMono: require("../assets/fonts/SpaceMono-Regular.ttf"),
@@ -33,26 +28,6 @@ export default function RootLayout() {
       SplashScreen.hideAsync();
     }
   }, [loaded]);
-  const { setUser } = useUser();
-
-  //TODO: Splash screen?
-  useEffect(() => {
-    async function fetchData() {
-      const token = await getToken();
-      if (token) {
-        const isValid = await verify();
-        if (isValid) {
-          const theUser: User = await getUser();
-          console.log("SETTING THE USER!", theUser);
-          setUser(theUser);
-          router.push("/(tabs)");
-        } else {
-          await deleteToken();
-        }
-      }
-    }
-    fetchData();
-  }, [router, setUser]);
 
   if (!loaded) {
     return null;
